requirements.txt
RPi.GPIO
python-vlc
schedule
dotenv
load_dotenv
schedule
nats-py
